whole app flow fix. (#17)
