+++ app/build.gradle
@@ -92,12 +92,12 @@ android {
             //versionCode project.inCode
             //versionName project.inVersion
 
-            buildConfigField 'String', 'CONCEPT', '"belagavi"'
+            buildConfigField 'String', 'CONCEPT', '"user"'
             buildConfigField 'String', 'BASE_URL', '"http://"'
             buildConfigField "Boolean", "SHOW_VERSION_TOAST", 'true'
             signingConfig signingConfigs.release
             //Release keys for one signal
-            manifestPlaceholders = [app_name                : "Pakr me",
+            manifestPlaceholders = [app_name                : "park me",
                                     ENABLE_CLEARTEXT_TRAFFIC: "false",
                                     google_maps_api_key     : ""
             ]
@@ -153,28 +153,20 @@ dependencies {
 
     // ViewModel
     implementation "androidx.lifecycle:lifecycle-extensions:$lifecycleVersion"
-    implementation 'androidx.lifecycle:lifecycle-common:2.5.0'
+    implementation 'androidx.lifecycle:lifecycle-common:2.5.1'
     //implementation 'androidx.lifecycle:lifecycle-viewmodel:2.5.0'
-    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.0'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1'
 
     // Kotlin Co-routines
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutineVersion"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutineVersion"
 
 
-    implementation 'com.google.code.gson:gson:2.8.9'
+    implementation 'com.google.code.gson:gson:2.9.0'
     // Retrofit
     implementation "com.jakewharton.retrofit:retrofit2-kotlin-coroutines-adapter:$retrofitCoroutinesVersion"
     implementation "com.squareup.retrofit2:converter-gson:$retrofitVersion"
     implementation "com.squareup.okhttp3:logging-interceptor:4.9.3"
-
-    //RxJava2
-     implementation 'io.reactivex.rxjava3:rxandroid:3.0.0'
-    implementation 'io.reactivex.rxjava3:rxandroid:3.0.0'
-    //implementation 'io.reactivex.rxjava3:rxjava:3.0.0'
-    //implementation "io.reactivex.rxjava2:rxandroid:2.0.1"
-    //implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
-    //implementation 'io.reactivex.rxjava2:rxjava:2.2.9'
 
     //google-Maps
     implementation "com.google.maps.android:android-maps-utils:$mapsUtilPlayServices"
@@ -194,7 +186,7 @@ dependencies {
 
     // FirebaseUI for Firebase Auth
     implementation 'com.firebaseui:firebase-ui-auth:7.1.1'
-    implementation 'com.google.firebase:firebase-auth-ktx:21.0.6'
+    implementation 'com.google.firebase:firebase-auth-ktx:21.0.7'
 
     //animation
     implementation 'com.airbnb.android:lottie:5.0.3'
