+++ app/src/main/java/com/anetos/parkme/core/helper/Networking.kt
@@ -1,11 +1,9 @@
 package  com.anetos.parkme.core
 
 import android.content.Context
+import com.anetos.parkme.BuildConfig
+import com.anetos.parkme.data.api.ApiService
 import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
-import  com.anetos.parkme.BuildConfig
-import  com.anetos.parkme.data.UrlConstants
-import  com.anetos.parkme.data.api.ApiService
-import  com.anetos.parkme.data.api.googleRemote.GoogleApiService
 import okhttp3.Cache
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
@@ -59,33 +57,5 @@ object Networking {
             .addCallAdapterFactory(CoroutineCallAdapterFactory())
             .build()
             .create(ApiService::class.java)
-    }
-
-    fun createGoogleApi(context: Context): GoogleApiService {
-        val cacheDir = File(context.cacheDir, "offlineCache")
-
-        //10 MB
-        val cacheSize = Cache(cacheDir, 10 * 1024 * 1024)
-
-        return Retrofit.Builder()
-            .baseUrl(UrlConstants.GOOGLE_BASE_URL)
-            .client(
-                OkHttpClient.Builder()
-                    .cache(cacheSize)
-                    .addInterceptor(HttpLoggingInterceptor()
-                        .apply {
-                            level = if (BuildConfig.DEBUG)
-                                HttpLoggingInterceptor.Level.BODY
-                            else
-                                HttpLoggingInterceptor.Level.NONE
-                        })
-                    .readTimeout(NETWORK_CALL_TIMEOUT.toLong(), TimeUnit.SECONDS)
-                    .writeTimeout(NETWORK_CALL_TIMEOUT.toLong(), TimeUnit.SECONDS)
-                    .build()
-            )
-            .addConverterFactory(GsonConverterFactory.create())
-            .addCallAdapterFactory(CoroutineCallAdapterFactory())
-            .build()
-            .create(GoogleApiService::class.java)
     }
 }
