+++ app/src/main/java/com/anetos/parkme/data/repository/Injection.kt
@@ -1,9 +1,8 @@
 package com.anetos.parkme.data.repository
 
 import android.content.Context
-import  com.anetos.parkme.data.api.googleRemote.GoogleRemoteDataSourceImpl
-import  com.anetos.parkme.data.api.remote.RemoteDataSourceImpl
-import  com.anetos.parkme.data.db.AppDatabase
+import com.anetos.parkme.data.api.remote.RemoteDataSourceImpl
+import com.anetos.parkme.data.db.AppDatabase
 
 /**
  * All the viewModel Injections will go here.
@@ -16,12 +15,5 @@ object Injection {
         DataRepository(
             RemoteDataSourceImpl.newInstance(context),
             AppDatabase.getInstance(context),
-            GoogleRemoteDataSourceImpl.newInstance(context)
-        )
-
-    fun provideGoogleDataRepository(context: Context) =
-        GoogleDataRepository(
-            GoogleRemoteDataSourceImpl.newInstance(context),
-            AppDatabase.getInstance(context)
         )
 }
