+++ app/src/main/res/drawable/bg_gradient_3.xml
@@ -5,4 +5,5 @@
         android:endColor="#3E3B92"
         android:startColor="#F44369"
         android:type="linear" />
+    <corners android:radius="@dimen/spacing_small" />
 </shape>
