+++ app/src/main/res/layout/dialog_loader.xml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!--
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:layout_margin="10dp"
-    android:background="@drawable/bg_black_solid_rect"
     android:paddingHorizontal="50dp"
     android:paddingVertical="15dp">
 
@@ -32,8 +30,11 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/progress_bar" />
 
-</androidx.constraintlayout.widget.ConstraintLayout>-->
-<com.airbnb.lottie.LottieAnimationView android:id="@+id/animation_view"
+</androidx.constraintlayout.widget.ConstraintLayout>
+<!--
+<com.airbnb.lottie.LottieAnimationView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/animation_view"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     app:layout_constraintBottom_toBottomOf="parent"
@@ -41,7 +42,5 @@
     app:layout_constraintStart_toStartOf="parent"
     app:layout_constraintTop_toTopOf="parent"
     app:lottie_autoPlay="true"
-    app:lottie_loop="true"
     app:lottie_fileName="ic_loading_circle_dots_white.json"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto" />
+    app:lottie_loop="true" />-->
