+++ app/src/main/res/values/themes.xml
@@ -215,15 +215,15 @@
     <style name="GradientTitle" parent="Widget.MaterialComponents.TextView">
         <item name="fontFamily">@font/nunito_extrabold</item>
         <item name="android:textColor">@color/white</item>
-        <item name="android:textSize">@dimen/size_extra_small</item>
+        <item name="android:textSize">@dimen/size_small</item>
         <item name="android:gravity">center</item>
     </style>
 
     <style name="GradientSubTitle" parent="Widget.MaterialComponents.TextView">
         <item name="fontFamily">@font/nunito_extrabold</item>
         <item name="android:textColor">@color/white</item>
-        <item name="android:textSize">@dimen/size_small</item>
-        <item name="android:gravity">center</item>
+        <item name="android:textSize">@dimen/size_extra_small</item>
+        <item name="android:gravity">start</item>
     </style>
 
     <style name="DialogItem" parent="Widget.MaterialComponents.TextView">
